+++ python-package/xgboost/_data_utils.py
@@ -745,82 +745,3 @@ class TransformedDf(ABC):
     @abstractmethod
     def shape(self) -> Tuple[int, int]:
         """Return the shape of the dataframe."""
-
-
-def _is_polars_lazyframe(data: DataType) -> bool:
-    return lazy_isinstance(data, "polars.lazyframe.frame", "LazyFrame")
-
-
-def _is_polars_series(data: DataType) -> bool:
-    return lazy_isinstance(data, "polars.series.series", "Series")
-
-
-def _is_polars(data: DataType) -> bool:
-    lf = _is_polars_lazyframe(data)
-    df = lazy_isinstance(data, "polars.dataframe.frame", "DataFrame")
-    return lf or df
-
-
-def _is_arrow(data: DataType) -> TypeGuard["pa.Table"]:
-    return lazy_isinstance(data, "pyarrow.lib", "Table")
-
-
-def _is_cudf_df(data: DataType) -> bool:
-    return lazy_isinstance(data, "cudf.core.dataframe", "DataFrame")
-
-
-def _is_cudf_ser(data: DataType) -> bool:
-    return lazy_isinstance(data, "cudf.core.series", "Series")
-
-
-def _is_cudf_pandas(data: DataType) -> bool:
-    """Must go before both pandas and cudf checks."""
-    return (_is_pandas_df(data) or _is_pandas_series(data)) and lazy_isinstance(
-        type(data), "cudf.pandas.fast_slow_proxy", "_FastSlowProxyMeta"
-    )
-
-
-def _is_pandas_df(data: DataType) -> TypeGuard["pd.DataFrame"]:
-    return lazy_isinstance(data, "pandas.core.frame", "DataFrame")
-
-
-def _is_pandas_series(data: DataType) -> TypeGuard["pd.Series"]:
-    return lazy_isinstance(data, "pandas.core.series", "Series")
-
-
-def _is_modin_df(data: DataType) -> bool:
-    return lazy_isinstance(data, "modin.pandas.dataframe", "DataFrame")
-
-
-def _is_modin_series(data: DataType) -> bool:
-    return lazy_isinstance(data, "modin.pandas.series", "Series")
-
-
-def is_dataframe(data: DataType) -> bool:
-    """Whether the input is a dataframe. Currently supported dataframes:
-
-    - pandas
-    - cudf
-    - cudf.pandas
-    - polars
-    - pyarrow
-    - modin
-
-
-    """
-    return any(
-        p(data)
-        for p in (
-            _is_polars_lazyframe,
-            _is_polars,
-            _is_polars_series,
-            _is_arrow,
-            _is_cudf_df,
-            _is_cudf_ser,
-            _is_cudf_pandas,
-            _is_pandas_df,
-            _is_pandas_series,
-            _is_modin_df,
-            _is_modin_series,
-        )
-    )
