+++ tests/python/test_with_pandas.py
@@ -5,7 +5,7 @@ import pytest
 
 import xgboost as xgb
 from xgboost import testing as tm
-from xgboost._data_utils import is_dataframe
+from xgboost.compat import is_dataframe
 from xgboost.core import DataSplitMode
 from xgboost.testing.data import pd_arrow_dtypes, pd_dtypes, run_base_margin_info
 from xgboost.testing.utils import predictor_equal
