+++ tests/python/test_with_polars.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 
 import xgboost as xgb
-from xgboost._data_utils import is_dataframe
+from xgboost.compat import is_dataframe
 
 pl = pytest.importorskip("polars")
